feat(cookbook): generic recipe for account creation
